feat: refactoring doc-tokens in new file and add new export function to table
